model_training/requirements.txt
tensorflow==2.15.0
numpy==1.26.2
pandas==2.1.3
matplotlib==3.8.2
pillow==10.1.0
scikit-learn==1.3.2
yfinance==0.2.28
mplfinance==0.12.10b0
opencv-python==4.8.1.78

